Begin to merge all key and cert management to a single module
